\name{iNextPD-package}
\alias{iNextPD-package}
\docType{package}
\title{
iNterpolation and EXTrapolation for phylogenetic diversity
}
\description{
iNterpolation and EXTrapolation for phylogenetic diversity
}
\details{
\tabular{ll}{
Package: \tab iNextPD\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2015-6-16\cr
License: \tab GPL (>= 3)\cr
}

}
\author{
T. C. Hsieh [aut, cre, cph]\cr
Chun-Huo Chiu [ctb]\cr
Anne Chao [ths]\cr
Maintainer: T. C. Hsieh <euler96@gmail.com>
}
\references{
Chao, Anne, et al. "Rarefaction and extrapolation of phylogenetic diversity." Methods in Ecology and Evolution 6.4 (2015): 380-388.
}
