% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{print.iNextPD}
\alias{print.iNextPD}
\title{Printing iNextPD object}
\usage{
\method{print}{iNextPD}(x, ...)
}
\arguments{
\item{x}{an \code{iNextPD} object computed by \code{\link{iNextPD}}.}

\item{...}{additional arguments.}
}
\description{
\code{print.iNextPD}: Print method for objects inheriting from class "iNextPD"
}

