\name{bird}
\alias{bird}
\docType{data}
\title{
Bird data in New South Wales
}
\description{
Species abundance frequency counts (abun), species names (lables) and phylogenetic tree (phy) data for two sites (South-site and North-site) at Barrington Tops National Park, Australia.
}
\usage{data(ant)}
\format{
}
\references{
Jetz, W., Thomas, G.H., Joy, J.B., Hartmann, K. & Mooers A.O. (2012). The global diversity of birds in space and time. Nature, 491, 444–448.
}

\keyword{datasets}
